normal, 1219

+++ mail/attachment/run.sh
@@ -10,13 +10,35 @@ if [ -z $token ]; then
 	exit
 fi
 
+py_version=`python -c 'import sys; print(sys.version_info.major)'`
+if [[ py_version -eq 3 ]]; then
+	py_command="python"
+fi
+if [[ py_version -eq 2 ]]; then
+	py_version=`python3 -c 'import sys; print(sys.version_info.major)'`
+	py_command="python3"
+fi
+if [[ -z $py_version ]]; then
+	echo "Please install latest python 3."
+	exit
+fi
+# echo $py_version
 GET "https://umjicanvas.com/api/v1/users/self/favorites/courses?access_token=$token" > /tmp/temp.txt
-courses=($(python ./canvas.py xxx -f -d $path | tr -d '[],'))
+courses=($($py_command ./canvas.py xxx -f -d $path | tr -d '[],'))
 rm /tmp/temp.txt
 # echo ${#courses[@]}
 len_course=${#courses[@]}
+# echo ${courses[@]}
+if [[ ${courses[0]} -eq -1 ]]; then
+	echo "ID not exists, or something went wrong with your token, please check."
+	exit
+fi
+if [[ ${courses[1]} -eq -1 ]]; then
+	echo "You do not have any course this semester, Bye!"
+	exit
+fi
 
-python ./canvas.py xxx -c -d $path
+$py_command ./canvas.py xxx -c -d $path
 for (( i = 1; i < ${len_course}; i += 2 ));
 do
 	echo "Fetching data from ${courses[$(($i + 1))]}"
@@ -32,7 +54,7 @@ do
 			break
 		fi
 	done
-	python ./canvas.py $course_name -d $path
+	$py_command ./canvas.py $course_name -d $path
 	rm /tmp/$course_name.txt
 done
 
@@ -40,9 +62,9 @@ my_activity=0
 for (( i = 1; i < ${len_course}; i += 2 ));
 do
 	course_name=${courses[$(($i + 1))]}
-	delta=`python ./canvas.py $course_name -m ${courses[0]} -d $path`
+	delta=`$py_command ./canvas.py $course_name -m ${courses[0]} -d $path`
 	my_activity=$((my_activity + $delta))
-	python ./canvas.py $course_name -r -d $path
+	$py_command ./canvas.py $course_name -r -d $path
 done
-python ./canvas.py xxx -s $my_activity -d $path
-python ./canvas.py xxx -p -d $path
+$py_command ./canvas.py xxx -s $my_activity -d $path
+$py_command ./canvas.py xxx -p -d $path
